refactoring changes for getRequest only
