+++ package.json
@@ -47,6 +47,7 @@
     "request": "2.88.0",
     "semver": "5.5.0",
     "superagent": "3.8.2",
+    "superagent-proxy": "^2.0.0",
     "url": "0.11.0",
     "uuid": "3.2.1",
     "yeoman-environment": "2.0.5"
